+++ pkg/db/db_test.go
@@ -58,36 +58,37 @@ func createMigrationFolder(dbLocation string) (string, error) {
 	return loc, os.Mkdir(loc, os.ModePerm)
 }
 
-func TestConnection(t *testing.T) {
-	tcs := []struct {
-		Name string
-	}{
-		{
-			Name: "Ping DB",
-		},
-	}
-	for _, tc := range tcs {
-		tc := tc
-		t.Run(tc.Name, func(t *testing.T) {
-			ctx := context.Background()
-			t.Parallel()
-			dir := t.TempDir()
-			cfg := DBConfig{
-				DriverName: "sqlite3",
-				DbHost:     dir,
-			}
-			db, err := Connect(ctx, cfg)
-			if err != nil {
-				t.Fatalf("Error establishing DB connection. Error: %v\n", err)
-			}
-			defer db.DB.Close()
-			pingErr := db.DB.Ping()
-			if pingErr != nil {
-				t.Fatalf("Error DB. Error: %v\n", err)
-			}
-		})
-	}
-}
+//
+//func TestConnection(t *testing.T) {
+//	tcs := []struct {
+//		Name string
+//	}{
+//		{
+//			Name: "Ping DB",
+//		},
+//	}
+//	for _, tc := range tcs {
+//		tc := tc
+//		t.Run(tc.Name, func(t *testing.T) {
+//			ctx := context.Background()
+//			t.Parallel()
+//			dir := t.TempDir()
+//			cfg := DBConfig{
+//				DriverName: "postgres",
+//				DbHost:     dir,
+//			}
+//			db, err := Connect(ctx, cfg)
+//			if err != nil {
+//				t.Fatalf("Error establishing DB connection. Error: %v\n", err)
+//			}
+//			defer db.DB.Close()
+//			pingErr := db.DB.Ping()
+//			if pingErr != nil {
+//				t.Fatalf("Error DB. Error: %v\n", err)
+//			}
+//		})
+//	}
+//}
 
 func TestMigrationScript(t *testing.T) {
 	tcs := []struct {
@@ -117,11 +118,12 @@ INSERT INTO apps (created, appname, statechange, metadata)  VALUES ('1713218400'
 			t.Parallel()
 			ctx := context.Background()
 			dbDir := t.TempDir()
-			cfg := DBConfig{
-				DriverName:     "sqlite3",
-				DbHost:         dbDir,
-				MigrationsPath: dbDir + "/migrations",
-			}
+			_, dbConfig := SetupRepositoryTestWithDBMigrationPath(t, dbDir)
+			//cfg := DBConfig{
+			//	DriverName:     "sqlite3",
+			//	DbHost:         dbDir,
+			//	MigrationsPath: dbDir + "/migrations",
+			//}
 			loc, mkdirErr := createMigrationFolder(dbDir)
 			if mkdirErr != nil {
 				t.Fatalf("Error creating migrations folder. Error: %v\n", mkdirErr)
@@ -134,12 +136,12 @@ INSERT INTO apps (created, appname, statechange, metadata)  VALUES ('1713218400'
 				t.Fatalf("Error creating migration file. Error: %v\n", mkdirErr)
 			}
 
-			migErr := RunDBMigrations(ctx, cfg)
+			migErr := RunDBMigrations(ctx, *dbConfig)
 			if migErr != nil {
 				t.Fatalf("Error running migration script. Error: %v\n", migErr)
 			}
 
-			db, err := Connect(ctx, cfg)
+			db, err := Connect(ctx, *dbConfig)
 			if err != nil {
 				t.Fatal("Error establishing DB connection: ", zap.Error(err))
 			}
@@ -268,7 +270,7 @@ func TestCustomMigrationReleases(t *testing.T) {
 			t.Parallel()
 			ctx := context.Background()
 
-			dbHandler := SetupRepositoryTestWithDB(t)
+			dbHandler, _ := SetupRepositoryTestWithDB(t)
 			err3 := dbHandler.WithTransaction(ctx, false, func(ctx context.Context, transaction *sql.Tx) error {
 				err2 := dbHandler.RunCustomMigrationReleases(ctx, getAllApps, writeAllReleases)
 				if err2 != nil {
@@ -338,7 +340,7 @@ func TestCustomMigrationsApps(t *testing.T) {
 			t.Parallel()
 			ctx := context.Background()
 
-			dbHandler := SetupRepositoryTestWithDB(t)
+			dbHandler, _ := SetupRepositoryTestWithDB(t)
 			err3 := dbHandler.WithTransaction(ctx, false, func(ctx context.Context, transaction *sql.Tx) error {
 				err2 := dbHandler.RunAllCustomMigrationsForApps(ctx, tc.allAppsFunc)
 				if err2 != nil {
@@ -390,7 +392,7 @@ func TestMigrationCommitEvent(t *testing.T) {
 			t.Parallel()
 			ctx := context.Background()
 
-			dbHandler := SetupRepositoryTestWithDB(t)
+			dbHandler, _ := SetupRepositoryTestWithDB(t)
 			err3 := dbHandler.WithTransaction(ctx, false, func(ctx context.Context, transaction *sql.Tx) error {
 				err2 := dbHandler.RunCustomMigrationsEventSourcingLight(ctx)
 				if err2 != nil {
@@ -499,7 +501,7 @@ func TestCommitEvents(t *testing.T) {
 			ctx := context.Background()
 			dbDir := t.TempDir()
 
-			dir, err := testutil.CreateMigrationsPath(2)
+			dir, err := CreateMigrationsPath(2)
 			if err != nil {
 				t.Fatalf("setup error could not detect dir \n%v", err)
 				return
@@ -649,7 +651,7 @@ func TestReadLockPreventedEvents(t *testing.T) {
 			ctx := context.Background()
 			dbDir := t.TempDir()
 
-			dir, err := testutil.CreateMigrationsPath(2)
+			dir, err := CreateMigrationsPath(2)
 			if err != nil {
 				t.Fatalf("setup error could not detect dir \n%v", err)
 				return
@@ -3175,7 +3177,7 @@ func TestReadWriteEslEvent(t *testing.T) {
 			t.Parallel()
 			ctx := testutil.MakeTestContext()
 			dbHandler := setupDB(t)
-			err := dbHandler.WithTransaction(ctx, true, func(ctx context.Context, transaction *sql.Tx) error {
+			err := dbHandler.WithTransaction(ctx, false, func(ctx context.Context, transaction *sql.Tx) error {
 				err := dbHandler.DBWriteEslEventInternal(ctx, tc.EventType, transaction, tc.EventData, tc.EventMetadata)
 				if err != nil {
 					return err
@@ -3310,7 +3312,7 @@ func TestReadWriteFailedEslEvent(t *testing.T) {
 			t.Parallel()
 			ctx := testutil.MakeTestContext()
 			dbHandler := setupDB(t)
-			err := dbHandler.WithTransaction(ctx, true, func(ctx context.Context, transaction *sql.Tx) error {
+			err := dbHandler.WithTransaction(ctx, false, func(ctx context.Context, transaction *sql.Tx) error {
 				err := dbHandler.DBWriteMigrationsTransformer(ctx, transaction)
 				if err != nil {
 					return err
@@ -4665,12 +4667,12 @@ func TestFindEnvAppsFromReleases(t *testing.T) {
 // setupDB returns a new DBHandler with a tmp directory every time, so tests can are completely independent
 func setupDB(t *testing.T) *DBHandler {
 	ctx := context.Background()
-	dir, err := testutil.CreateMigrationsPath(2)
+	dir, err := CreateMigrationsPath(2)
 	tmpDir := t.TempDir()
 	t.Logf("directory for DB migrations: %s", dir)
 	t.Logf("tmp dir for DB data: %s", tmpDir)
 
-	dbConfig, err := testutil.SetupPostgresContainer(t, dir, false)
+	dbConfig, err := SetupPostgresContainer(ctx, t, dir, false, t.Name())
 	if err != nil {
 		t.Fatalf("SetupPostgres: %v", err)
 	}
@@ -4686,7 +4688,7 @@ func setupDB(t *testing.T) *DBHandler {
 		t.Fatal(migErr)
 	}
 
-	dbHandler, err := Connect(ctx, cfg)
+	dbHandler, err := Connect(ctx, *dbConfig)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -4694,16 +4696,21 @@ func setupDB(t *testing.T) *DBHandler {
 	return dbHandler
 }
 
-func SetupRepositoryTestWithDB(t *testing.T) *DBHandler {
-	ctx := context.Background()
-	migrationsPath, err := testutil.CreateMigrationsPath(2)
+func SetupRepositoryTestWithDB(t *testing.T) (*DBHandler, *DBConfig) {
+	migrationsPath, err := CreateMigrationsPath(2)
 	if err != nil {
 		t.Fatalf("CreateMigrationsPath error: %v", err)
 	}
-	dbConfig := &DBConfig{
-		MigrationsPath: migrationsPath,
-		DriverName:     "sqlite3",
-	}
+	return SetupRepositoryTestWithDBMigrationPath(t, migrationsPath)
+}
+
+func SetupRepositoryTestWithDBMigrationPath(t *testing.T, migrationsPath string) (*DBHandler, *DBConfig) {
+	ctx := context.Background()
+	dbConfig, err := SetupPostgresContainer(ctx, t, migrationsPath, false, t.Name())
+	//dbConfig := &DBConfig{
+	//	MigrationsPath: migrationsPath,
+	//	DriverName:     "sqlite3",
+	//}
 
 	dir := t.TempDir()
 	remoteDir := path.Join(dir, "remote")
@@ -4712,27 +4719,27 @@ func SetupRepositoryTestWithDB(t *testing.T) *DBHandler {
 	err = cmd.Start()
 	if err != nil {
 		t.Fatalf("error starting %v", err)
-		return nil
+		return nil, nil
 	}
 	err = cmd.Wait()
 	if err != nil {
 		t.Fatalf("error waiting %v", err)
-		return nil
+		return nil, nil
 	}
 	t.Logf("test created dir: %s", localDir)
 
-	dbConfig.DbHost = dir
+	//dbConfig.DbHost = dir
 
-	migErr := RunDBMigrations(ctx, *dbConfig)
-	if migErr != nil {
-		t.Fatal(migErr)
-	}
+	//migErr := RunDBMigrations(ctx, *dbConfig)
+	//if migErr != nil {
+	//	t.Fatal(migErr)
+	//}
 
 	dbHandler, err := Connect(ctx, *dbConfig)
 	if err != nil {
 		t.Fatal(err)
 	}
-	return dbHandler
+	return dbHandler, dbConfig
 }
 
 func TestReadReleasesWithoutEnvironments(t *testing.T) {
