+++ services/cd-service/pkg/repository/repository_test.go
@@ -213,7 +213,7 @@ func TestApplyQueuePanic(t *testing.T) {
 			t.Parallel()
 			// create a remote
 			ctx := testutil.MakeTestContext()
-			migrationsPath, err := testutil.CreateMigrationsPath(4)
+			migrationsPath, err := db.CreateMigrationsPath(4)
 			if err != nil {
 				t.Fatalf("CreateMigrationsPath error: %v", err)
 			}
@@ -308,7 +308,7 @@ func TestApplyQueueTtlForHealth(t *testing.T) {
 		tc := tc
 		t.Run(tc.Name, func(t *testing.T) {
 			ctx, cancel := context.WithTimeout(testutil.MakeTestContext(), 10*time.Second)
-			migrationsPath, err := testutil.CreateMigrationsPath(4)
+			migrationsPath, err := db.CreateMigrationsPath(4)
 			if err != nil {
 				t.Fatalf("CreateMigrationsPath error: %v", err)
 			}
@@ -1110,7 +1110,7 @@ func TestMeasureGitSyncStatus(t *testing.T) {
 
 func SetupRepositoryBenchmark(t *testing.B, writeEslOnly bool) (Repository, *db.DBHandler) {
 	ctx := context.Background()
-	migrationsPath, err := testutil.CreateMigrationsPath(4)
+	migrationsPath, err := db.CreateMigrationsPath(4)
 	if err != nil {
 		t.Fatalf("CreateMigrationsPath error: %v", err)
 	}
