+++ services/cd-service/pkg/repository/transformer_test.go
@@ -3313,7 +3313,7 @@ func SetupRepositoryTestWithDB(t *testing.T) Repository {
 
 func SetupRepositoryTestWithDBOptions(t *testing.T, writeEslOnly bool) (Repository, *db.DBHandler) {
 	ctx := context.Background()
-	migrationsPath, err := testutil.CreateMigrationsPath(4)
+	migrationsPath, err := db.CreateMigrationsPath(4)
 	if err != nil {
 		t.Fatalf("CreateMigrationsPath error: %v", err)
 	}
