+++ services/manifest-repo-export-service/pkg/service/version_test.go
@@ -43,7 +43,7 @@ import (
 
 func setupRepositoryTestWithPath(t *testing.T) (repository.Repository, string) {
 	ctx := context.Background()
-	migrationsPath, err := testutil.CreateMigrationsPath(4)
+	migrationsPath, err := db.CreateMigrationsPath(4)
 	if err != nil {
 		t.Fatalf("CreateMigrationsPath error: %v", err)
 	}
