+++ services/rollout-service/pkg/versions/versions_test.go
@@ -30,7 +30,6 @@ import (
 	"github.com/freiheit-com/kuberpult/pkg/config"
 	"github.com/freiheit-com/kuberpult/pkg/db"
 	"github.com/freiheit-com/kuberpult/pkg/setup"
-	"github.com/freiheit-com/kuberpult/pkg/testutil"
 	"github.com/google/go-cmp/cmp"
 	grpc "google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -952,7 +951,7 @@ func assertExpectedVersions(t *testing.T, expectedVersions []expectedVersion, vc
 // setupDB returns a new DBHandler with a tmp directory every time, so tests can are completely independent
 func setupDB(t *testing.T) *db.DBHandler {
 	ctx := context.Background()
-	dir, err := testutil.CreateMigrationsPath(2)
+	dir, err := db.CreateMigrationsPath(2)
 	tmpDir := t.TempDir()
 	t.Logf("directory for DB migrations: %s", dir)
 	t.Logf("tmp dir for DB data: %s", tmpDir)
